Move getPlaceholder to utils
